requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
pytorch-lightning>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
redis>=4.5.0
faiss-cpu>=1.7.4
transformers>=4.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
praw>=7.7.0
python-dotenv>=1.0.0
tqdm>=4.65.0
networkx>=3.1
matplotlib>=3.7.0
plotly>=5.15.0
dash>=2.11.0 